requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
discord.py
openai
pydantic
python-multipart
typing-extensions